+++ tests/pna/config/test_config.py
@@ -314,10 +314,13 @@ def test_list_panel_names(pna_data_root):
             "proxiome-v1-immuno-156-FMC63-v1.1",
             "proxiome-v2-immuno-155-v1.0",
             "proxiome-v2-immuno-155-v2.0",
+            "proxiome-v2-immuno-155-v2.1",
             "proxiome-v2-immuno-156-FLAG-v1.0",
             "proxiome-v2-immuno-156-FLAG-v2.0",
+            "proxiome-v2-immuno-156-FLAG-v2.1",
             "proxiome-v2-immuno-156-FMC63-v1.0",
             "proxiome-v2-immuno-156-FMC63-v2.0",
+            "proxiome-v2-immuno-156-FMC63-v2.1",
         ]
     )
 
@@ -329,8 +332,11 @@ def test_list_panel_names(pna_data_root):
         "proxiome-v1-immuno-156-FMC63-v1.0",
         "proxiome-v1-immuno-156-FMC63-v1.1",
         "proxiome-v2-immuno-155-v2.0",
+        "proxiome-v2-immuno-155-v2.1",
         "proxiome-v2-immuno-156-FLAG-v2.0",
+        "proxiome-v2-immuno-156-FLAG-v2.1",
         "proxiome-v2-immuno-156-FMC63-v2.0",
+        "proxiome-v2-immuno-156-FMC63-v2.1",
     ]